+++ argilla-frontend/translation/es.js
@@ -22,6 +22,7 @@ export default {
   description: "Descripción",
   labels: "Etiquetas",
   order: "Orden",
+  owner: "Propietario",
   useMarkdown: "Usar Markdown",
   suggestionFirst: "Mostrar sugerencias primero",
   visibleForAnnotators: "Visible para los anotadores",
@@ -42,6 +43,7 @@ export default {
   optional: "Opcional",
   template: "Plantilla",
   rows: "filas",
+  datasetName: "Dataset name",
   noRecordsMessages: {
     datasetEmptyForAnnotator:
       "El dataset está vacío. Pide a un administrador que suba registros y vuelve pronto.",
@@ -326,6 +328,23 @@ export default {
     none: "Ninguno",
     noWorkspaces:
       "Por favor, sigue <a target='_blank' href='https://docs.argilla.io/latest/how_to_guides/workspace/#create-a-new-workspace'>esta guía</a> para crear un espacio de trabajo",
+  },
+  exportToHub: {
+    dialogTitle: "Exportar dataset a Hugging Face",
+    ownerTooltip:
+      "Utiliza un nombre de usuario u organización de Hugging Face válidos",
+    tokenTooltip: `Utiliza un token de acceso existente o crear un <a href='https://huggingface.co/settings/tokens' target='_blank'>nuevo token</a> con "permiso de escritura"`,
+    validations: {
+      orgOrUsernameIsRequired:
+        "El nombre de usuario u organización es requerido",
+      hfTokenIsRequired: "El token de Hugging Face es requerido",
+      hfTokenInvalid: "El token de Hugging Face es inválido",
+      datasetNameIsRequired: "El nombre del dataset es requerido",
+    },
+    exporting: "Exportando al hub de Hugging Face",
+    private: "dataset privado",
+    public: "dataset público",
+    exportingWarning: "Esto puede tardar unos segundos",
   },
   config: {
     field: {