+++ argilla-server/tests/unit/conftest.py
@@ -12,16 +12,14 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import uuid
-from typing import TYPE_CHECKING, Dict, Generator, Optional
-
 import pytest
 import pytest_asyncio
-from sqlalchemy.engine.interfaces import IsolationLevel
+
+from typing import TYPE_CHECKING, Dict, Generator, Optional
 from httpx import AsyncClient
 from opensearchpy import OpenSearch
+from sqlalchemy.engine.interfaces import IsolationLevel
 
-from argilla_server import telemetry
 from argilla_server.contexts import distribution, datasets, records
 from argilla_server.api.routes import api_v1
 from argilla_server.constants import API_KEY_HEADER_NAME, DEFAULT_API_KEY
@@ -30,6 +28,7 @@ from argilla_server.models import User, UserRole, Workspace
 from argilla_server.search_engine import SearchEngine, get_search_engine
 from argilla_server.settings import settings
 from argilla_server.telemetry import TelemetryClient
+
 from tests.database import TestSession
 from tests.factories import AnnotatorFactory, OwnerFactory, UserFactory
 
@@ -91,7 +90,6 @@ async def async_client(
 
     mocker.patch.object(distribution, "_get_async_db", override_get_async_db)
     mocker.patch.object(datasets, "get_async_db", override_get_async_db)
-    mocker.patch.object(records, "get_async_db", override_get_async_db)
 
     api_v1.dependency_overrides.update(
         {